+++ orthanc_api_client/instances_set.py
@@ -22,7 +22,12 @@ class InstancesSet:
         self.study_id = None
 
     def add_series(self, series_id: str):
-        instances_ids = self._api_client.series.get_instances_ids(orthanc_id=series_id)
+        self._add_series(
+            series_id=series_id,
+            instances_ids=self._api_client.series.get_instances_ids(orthanc_id=series_id)
+        )
+
+    def _add_series(self, series_id: str, instances_ids: List[str]):
         self._by_series[series_id] = instances_ids
         self._all_instances_ids.extend(instances_ids)
 
@@ -145,6 +150,8 @@ class InstancesSet:
     # this method returns the list of removed instances IDs
     def filter_instances(self, filter) -> List[str]:
         series_to_delete = []
+        filtered = InstancesSet(self._api_client)
+        filtered.study_id = self.study_id
 
         for series_id, instances_ids in self._by_series.items():
             instances_to_delete = []
@@ -159,10 +166,12 @@ class InstancesSet:
             if len(self._by_series[series_id]) == 0:
                 series_to_delete.append(series_id)
 
+            filtered._add_series(series_id, instances_to_delete)
+
         for s in series_to_delete:
             del self._by_series[s]
 
-        return instances_to_delete
+        return filtered
 
     # apply a method on all instances
     # prototype: processor(api_client, instance_id)